// types/job/report.ts
export enum ReportStatus {
  REPORT_STATUS_UNSPECIFIED = 0,
  PENDING = 1,
  UNDER_REVIEW = 2,
  RESOLVED = 3,
  REJECTED = 4,
  CLOSED = 5,
  UNRECOGNIZED = -1,
}

/** Message */
export interface Report {
  id: string;
  title?: string | undefined;
  reporterId: string;
  reportedUserId: string;
  reason?: string | undefined;
  status: ReportStatus;
  createdAt: string;
}

/**
 * ---------------------
 * CRUD Responses
 * ---------------------
 */
export interface ReportResponse {
  success: boolean;
  message?: string | undefined;
}

export interface ListReportsResponse {
  reports: Report[];
  total: number;
  page: number;
  limit: number;
}

export interface DeleteReportResponse {
  success: boolean;
  message: string;
}

export interface AllReportsWithMetaResponse {
  data: Report[];
  meta: PaginationMeta | undefined;
}

export interface SingleReportResponse {
  message?: string | undefined;
  report: Report | undefined;
}

/**
 * ---------------------
 * CRUD Requests
 * ---------------------
 */
export interface ListReportsRequest {
  page: number;
  limit: number;
  status?: ReportStatus | undefined;
  reporterId?: string | undefined;
  reportedUserId?: string | undefined;
}

export interface GetReportRequest {
  reportId: string;
}

export interface UpdateReportRequest {
  reportId: string;
  title?: string | undefined;
  reason?: string | undefined;
  status?: ReportStatus | undefined;
}

export interface DeleteReportRequest {
  reportId: string;
}

export interface GetAllReportsRequest {
  basicQuery: BasicQuery | undefined;
}

export interface CreateReportRequest {
  title?: string | undefined;
  reporterId: string;
  reportedUserId: string;
  reason?: string | undefined;
  status?: ReportStatus | undefined;
}